+++ web/api/exporters/allmetrics.h
@@ -3,7 +3,7 @@
 #ifndef NETDATA_API_ALLMETRICS_H
 #define NETDATA_API_ALLMETRICS_H
 
-#include "../rrd2json.h"
+#include "web/api/formatters/rrd2json.h"
 #include "shell/allmetrics_shell.h"
 
 extern int web_client_api_request_v1_allmetrics(RRDHOST *host, struct web_client *w, char *url);
