+++ web/api/formatters/rrd2json.h
@@ -3,9 +3,18 @@
 #ifndef NETDATA_RRD2JSON_H
 #define NETDATA_RRD2JSON_H 1
 
-#include "web_api_v1.h"
-#include "exporters/allmetrics.h"
-#include "queries/rrdr.h"
+#include "web/api/web_api_v1.h"
+#include "web/api/exporters/allmetrics.h"
+#include "web/api/queries/rrdr.h"
+
+#include "web/api/formatters/csv/csv.h"
+#include "web/api/formatters/ssv/ssv.h"
+#include "web/api/formatters/json/json.h"
+#include "web/api/formatters/value/value.h"
+
+#include "web/api/formatters/rrdset2json.h"
+#include "web/api/formatters/charts2json.h"
+#include "web/api/formatters/json_wrapper.h"
 
 #define HOSTNAME_MAX 1024
 
@@ -38,7 +47,7 @@
 #define DATASOURCE_FORMAT_CSV_JSON_ARRAY "csvjsonarray"
 
 extern void rrd_stats_api_v1_chart(RRDSET *st, BUFFER *wb);
-extern void rrd_stats_api_v1_charts(RRDHOST *host, BUFFER *wb);
+extern void rrdr_buffer_print_format(BUFFER *wb, uint32_t format);
 
 extern int rrdset2anything_api_v1(
           RRDSET *st
