+++ web/api/web_api_v1.c
@@ -224,7 +224,7 @@ inline int web_client_api_request_v1_charts(RRDHOST *host, struct web_client *w,
 
     buffer_flush(w->response.data);
     w->response.data->contenttype = CT_APPLICATION_JSON;
-    rrd_stats_api_v1_charts(host, w->response.data);
+    charts2json(host, w->response.data);
     return 200;
 }
